Setup standard for code formatting
While I do care about code formatting (a great deal), I don't care for arguments
about it. The way I see it, we have 3 options:

- [Default Rubocop rules](https://github.com/rubocop-hq/rubocop)
- [Standard](https://github.com/testdouble/standard)
- [Substance Lab Style](https://github.com/substancelab/dotfiles/blob/master/rubocop.yml)

While I (naturally) prefer the latter and use it anywhere that isn't an open
source project, the former is likely the least controversial. However, seeing
how we're trying to remove arguments, defaulting to standard puts us in a
position where there are no settings to argue over.

Thus, here we are.